fix Bug #72049, when mv changed, generate changed source org cycle tasks.
