+++ core/src/main/java/inetsoft/sree/RepletRegistry.java
@@ -17,6 +17,7 @@
  */
 package inetsoft.sree;
 
+import inetsoft.mv.MVTool;
 import inetsoft.sree.internal.*;
 import inetsoft.sree.security.*;
 import inetsoft.uql.XPrincipal;
@@ -499,7 +500,7 @@ public class RepletRegistry implements Serializable {
          listeners = new Vector<>(this.listeners);
       }
 
-      PropertyChangeEvent evt = new PropertyChangeEvent(src, name, oval, nval);
+      PropertyChangeEvent evt = new PropertyChangeEvent(src, MVTool.getOrgEventSourceID(name, orgID), oval, nval);
 
       fireEventToListeners(listeners, evt, name);
