+++ core/src/main/java/inetsoft/web/admin/content/repository/MVChangeController.java
@@ -18,6 +18,7 @@
 package inetsoft.web.admin.content.repository;
 
 import inetsoft.mv.MVManager;
+import inetsoft.mv.MVTool;
 import inetsoft.sree.internal.SUtil;
 import inetsoft.sree.internal.cluster.*;
 import inetsoft.sree.security.OrganizationManager;
@@ -66,7 +67,7 @@ public class MVChangeController implements MessageListener {
 
    private void mvPropertyChanged(PropertyChangeEvent event) {
       if(MVManager.MV_CHANGE_EVENT.equals(event.getPropertyName())) {
-         String orgId = MVManager.getOrgIdFromEventSource(event.getSource());
+         String orgId = MVTool.getOrgIdFromEventSource(event.getSource());
 
          if(orgId != null &&
             !Tool.equals(orgId, OrganizationManager.getInstance().getCurrentOrgID(principal)))
